+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v3.0.3
+#### 🚀 Features
+
+- Support Qwen 3.5 models (PR: #391)
+- Upgrade LlamaLib to v2.0.4 (llama.cpp b8209) (PR: #391)
+- Android GPU support with Vulkan (PR: #390)
+- add button to redownload LlamaLib (PR: #393)
+
+
 ## v3.0.2
 #### 🚀 Features